Unify some code between image blobs and instance path
